BookRecommendationSystem/backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.3
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
PyJWT==2.8.0
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.25.1
whitenoise==6.6.0
